pyproject.toml
[tool.poetry]
name = "rinha-de-backend-2023"
version = "0.1.0"
description = ""
authors = ["Alcides Ribeiro <alcidesribeirosampaioneto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.0"
taskipy = "^1.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
extend-exclude = "/(?!rinha_de_backend_2023|tests)[^/]+/"

[tool.isort]
profile = "black"
skip = [".venv", "migrations"]

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
format = "black . && isort ."
run = "uvicorn rinha_de_backend_2023.main:app --reload"
pre_test = "task lint"
test = "pytest -s -x --cov=stock_backend -vv"
post_test = "coverage html"
